replaced option position type to int.

+++ src/ArgsMapper/ArgsCommandSettingsExtensions.cs
@@ -126,7 +126,7 @@ namespace ArgsMapper
             bool isPositional, Action<ArgsOptionSettings<TOption>> optionSettings)
             where TCommand : class
         {
-            ushort? position = null;
+            int? position = null;
 
             if (isPositional)
             {
